CRE-283 Send Metering Report to Beholder (#16849)
* Send Metering Report to Beholder

This PR collects metering information from workflow steps, builds a report and sends it as a custom message to beholder.
Metering values are persisted along with the workflow.

The metering report is currently sent as a string value for `BaseMessage` as a temporary measure, but the proto definition
exists such that the metering report can be sent as a custom proto message when the infrastructure exists to do so.

* convert metering message send to test hook

* update tests for new metering report structure

* update metering report SetStep with better var name

* update common

* make linter happy

* fix error with workflow limiter