fix(operations): improve report type conversion (#17028)
Operations reports are read from disk and will be consumed during operation execution during migration.
There will be a point where we need to convert the type from Report[any,any] into Report[IN,OUT] due to the type information being known later during the execution.

However, the existing conversion method was incorrect as the type casting is not reliable enough since when json marshalling and unmarshalling, the type information is lost, struct is converted into maps and int becomes float64.
Eg when marshaling a struct and unmarhalling back it becomes a map[string]interface{}.

This commit addresses this by using json.unmarshal to perform the type conversion instead which is more reliable.

JIRA: https://smartcontract-it.atlassian.net/browse/CLD-50