CRE-293 Add Metering Detail to Capabilities Response (#16918)
* Add Metering Detail to Capabilities Response

Every node must independently report metering information while the engine expects an aggregated list of all nodes. The
client expects an encoded `CapabilityResponse` as the payload with exactly one metering report value. These values are
then aggregated for the expected number of response messages and included as metadata in the response forwarded to the
engine. This approach isolates the aggregation logic to a single location in the code.

* simplify function output

* code cleanup, linting, and go mod

* address feedback and fix CI

* adding with_metering_metadata test

* gauge --> counter; adding labels to logger

* removing counter increment when unecessary

---------

Co-authored-by: patrickhuie19 <patrick.huie@smartcontract.com>